+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+## [0.5.0][] (2021-01-06)
+
 New:
 
 - Replace virtus with dry-struct.
@@ -147,7 +149,8 @@ Fixes:
 
 - Initial release
 
-[Unreleased]: https://github.com/opti/panda_doc/compare/v0.4.3...HEAD
+[Unreleased]: https://github.com/opti/panda_doc/compare/v0.5.0...HEAD
+[0.5.0]: https://github.com/opti/panda_doc/compare/v0.4.3...v0.5.0
 [0.4.3]: https://github.com/opti/panda_doc/compare/v0.4.2...v0.4.3
 [0.4.2]: https://github.com/opti/panda_doc/compare/v0.4.1...v0.4.2
 [0.4.1]: https://github.com/opti/panda_doc/compare/v0.4.0...v0.4.1